+++ apps/examples/blinkid-core-api/CHANGELOG.md
@@ -1,5 +1,11 @@
 # example-blinkid-core-api
 
+## 1.1.4
+
+### Patch Changes
+
+- @microblink/blinkid-core@7.2.0
+
 ## 1.1.3
 
 ### Patch Changes