+++ packages/blinkid-wasm/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @microblink/blinkid-wasm
 
+## 7.2.0
+
+### Minor Changes
+
+- 3655cc6: Update WASM files
+- 5d08204: - Implemented `showDemoOverlay` and `showProductionOverlay`
+  - Various bugfixes
+
 ## 7.1.0
 
 ### Minor Changes